Enable stylua for all files
