+++ .styluaignore
@@ -1,19 +1,4 @@
-# exclude everything by default
-/*
-
-# opt-in new code through a series of exclusions and inclusions https://stackoverflow.com/questions/5533050/
-!.styluaignore
-!.stylua.toml
-!/Scripts
-/Scripts/*
-!/Scripts/DCS-BIOS
-/Scripts/DCS-BIOS/*
-!/Scripts/DCS-BIOS/lib
-/Scripts/DCS-BIOS/lib/*
-!/Scripts/DCS-BIOS/lib/modules
-!/Scripts/DCS-BIOS/lib/io
-!Scripts/DCS-BIOS/lib/common
-# opt-in test code
-!/Scripts/DCS-BIOS/test
+# opt-out archived code
+/Scripts/DCS-BIOS/lib/archive
 # opt-out copy-pasted test library code
 /Scripts/DCS-BIOS/test/ext
