+++ Scripts/DCS-BIOS/BIOSConfig.lua
@@ -18,7 +18,7 @@ local BIOSConfig = {
 	tcp_config = {
 		{
 			address = "*",
-			port = 7778
+			port = 7778,
 		},
 	},
 	udp_config = {
@@ -26,7 +26,7 @@ local BIOSConfig = {
 			send_address = "239.255.50.10",
 			send_port = 5010,
 			receive_address = "*",
-			receive_port = 7778
+			receive_port = 7778,
 		},
 	},
 	dev_mode = true,
