+++ Scripts/Export.lua
@@ -1,1 +1,1 @@
-dofile(lfs.writedir()..[[Scripts\DCS-BIOS\BIOS.lua]])
+dofile(lfs.writedir() .. [[Scripts\DCS-BIOS\BIOS.lua]])
